requirements.txt
# AI Bug Bounty Scanner Dependencies - Enhanced Version 2.0

# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3

# Environment configuration
python-dotenv==1.0.0

# Database dependencies
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.12.1

# Async task processing
celery==5.3.4
redis==5.0.1
flower==2.0.1
kombu==5.3.4

# Authentication and security
PyJWT==2.8.0
Werkzeug==2.3.7
cryptography==41.0.7
bcrypt==4.0.1

# Security scanning libraries
python-nmap==0.7.1
requests==2.31.0
dnspython==2.4.2
python-whois==0.8.0
shodan==1.31.0

# Threat intelligence libraries
aiohttp==3.9.1
asyncio

# External security tools integration
# Note: These tools need to be installed separately
# sublist3r - Clone from GitHub
# amass - Download binary from GitHub releases
# nuclei - Download binary from GitHub releases  
# sqlmap - Clone from GitHub

# HTML parsing and web scraping
beautifulsoup4==4.12.2
lxml==4.9.3
fake-useragent==1.4.0

# Network and URL utilities
urllib3==2.0.7
validators==0.22.0
tldextract==5.1.1

# Async support
aiohttp==3.9.1

# Data processing and analysis
pandas==2.1.4
numpy==1.24.4
python-dateutil==2.8.2

# Report generation
reportlab==4.0.5
markdown==3.5.1
Jinja2==3.1.2

# File handling
python-magic==0.4.27

# Configuration and environment
python-dotenv==1.0.0
configparser==6.0.0

# Monitoring and deployment
gunicorn==21.2.0
supervisor==4.2.5

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.9.1
flake8==6.1.0
coverage==7.3.2

# Logging and debugging
colorlog==6.7.0
reportlab==4.0.7
weasyprint==60.2
jinja2==3.1.2
markdown==3.5.1

# Authentication and security
bcrypt==4.1.2
passlib==1.7.4
python-jose==3.3.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# API and HTTP utilities
httpx==0.25.2
python-multipart==0.0.6

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0
